Moved tests into test folder
