--- .babelrc
@@ -1,13 +1,0 @@
-{
-  "env": {
-    "test": {
-      "presets": [
-        "@babel/preset-env",
-        "@babel/preset-react",
-        "@babel/preset-typescript@7.15.0"
-      ],
-      "plugins": ["transform-export-extensions"],
-      "only": ["./**/*.ts", "./**/*.tsx", "node_modules/jest-runtime"]
-    }
-  }
-}
