+++ .github/workflows/main.yml
@@ -14,6 +14,6 @@ jobs:
       - uses: borales/actions-yarn@v2.3.0
         with:
           cmd: install # will run `yarn install` command
-      - uses: Borales/actions-yarn@v2.3.0
+      - uses: borales/actions-yarn@v2.3.0
         with:
           cmd: test # will run `yarn test` command
