+++ src/__tests__/features/Footer/Footer.test.tsx
@@ -1,4 +1,4 @@
-import Footer from './Footer'
+import Footer from '../../../features/Footer/Footer'
 
 import { render, screen } from '@testing-library/react'
