+++ src/features/LoadingPage/LoadingPage.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react'
 
-import '../../styles/LoadingPage.scss'
+import 'styles/LoadingPage.scss'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faSpinner } from '@fortawesome/free-solid-svg-icons'
