+++ src/setUpTests.ts
@@ -1,2 +1,4 @@
 import '@testing-library/jest-dom/extend-expect'
 import '@testing-library/jest-dom'
+
+export {}
